Leaf 265: split OK preamble consumers module
